code/requirements.txt
speedtest-cli
sqlalchemy
psycopg2-binary
fpdf2
matplotlib